+++ spec/default_values_spec.rb
@@ -54,12 +54,12 @@ describe "default values" do
       class Test::Bar
         extend Dry::Initializer::Mixin
 
-        param  :foo
-        param  :bar, default: proc { {} }
+        param :foo
+        param :bar, default: proc { {} }
       end
     end
 
-    it "instantiate arguments" do
+    it "instantiates arguments" do
       subject = Test::Bar.new(1, 2)
 
       expect(subject.foo).to eql 1
@@ -73,7 +73,7 @@ describe "default values" do
       expect(subject.bar).to eql({})
     end
 
-    it "instantiate arguments also if the last is an hash" do
+    it "instantiates arguments also if the last is an hash" do
       subject = Test::Bar.new(1, { baz: 2, qux: 3 })
 
       expect(subject.foo).to eql 1