+++ spec/missed_default_spec.rb
@@ -9,6 +9,6 @@ describe "missed default values" do
   end
 
   it "raises SyntaxError" do
-    expect { subject }.to raise_error RuntimeError, /bar/
+    expect { subject }.to raise_error ArgumentError, /bar/
   end
 end